services/incident-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
kafka-python==2.0.2
pydantic-settings==2.8.1
prometheus-client==0.22.1
